smart_traffic/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.3.0
gunicorn